+++ packages/xchain-thorchain/src/index.ts
@@ -1,3 +1,4 @@
 export * from './client'
+export * from './messages'
 export * from './types'
 export * from './util'
